fix: cast initial_basis to long in 07_iterative_nqs_dci scripts
`FlowGuidedKrylovPipeline.extract_and_select_basis()` returns a float32
tensor (cloned from the NF trainer's `accumulated_basis`, which lives in
float for gradient tracking). But `run_hi_nqs_sqd(initial_basis=...)` and
`run_hi_nqs_skqd(initial_basis=...)` strictly validate that
`initial_basis` is integer/bool dtype (binary occupation vectors) and
raise ValueError on float.

This caused both `iter_nqs_dci_sqd.py` and `iter_nqs_dci_krylov_classical.py`
to fail end-to-end with:

    ValueError: initial_basis must be integer or bool dtype
    (binary occupations), got torch.float32

at the `run_hi_nqs_sqd`/`run_hi_nqs_skqd` call site.

Interestingly, the parallel scripts in group 08 (iter_nqs_dci_pt2_*) work
because their basis is passed through `expand_basis_via_connections()`
first, which happens to cast internally. The 007 scripts do not have that
intermediate step, so they hit the bug raw.

Fix: add `.long()` cast to the `basis` result in the two affected scripts,
with a short comment explaining why. Both scripts now reach chemical
accuracy on H2:

    07_iterative_nqs_dci/iter_nqs_dci_sqd.py           h2 cpu 11.9s  0.0 mHa
    07_iterative_nqs_dci/iter_nqs_dci_krylov_classical h2 cpu  9.3s  0.0 mHa

Scope:
- Surgical script-level fix (2 files, 2-line changes each including
  comment). Does not change library-level API contracts.
- The 3rd script in the group (`iter_nqs_dci_krylov_quantum.py`) does not
  use `initial_basis` at all — unaffected.
- Related architectural issue (float32 basis at API boundary with int-only
  validator) could eventually be fixed by either (a) making
  `extract_and_select_basis` return long, or (b) making
  `validate_initial_basis` lenient for float {0,1} values. Out of scope
  for this minimal fix.

Discovered via end-to-end smoke test on `refactor/pipeline-catalog` branch
(PR #39). Verified that the failure is pre-existing on main before writing
this fix.

+++ experiments/pipelines/07_iterative_nqs_dci/iter_nqs_dci_sqd.py
@@ -159,7 +159,11 @@ def main() -> None:
 
     history = pipeline.train_flow_nqs(progress=True)
     n_epochs = len(history.get("total_loss", []))
-    basis = pipeline.extract_and_select_basis()
+    # `extract_and_select_basis()` returns float32 (inherited from the
+    # NF trainer's accumulated_basis). `run_hi_nqs_sqd(initial_basis=...)`
+    # requires integer/bool dtype for the binary occupation vectors, so
+    # cast before passing.
+    basis = pipeline.extract_and_select_basis().long()
     nf_dci_energy = pipeline.results.get("final_energy")
     nf_time = time.perf_counter() - t_start
